// at-common/model/ReportRevenueCriterial.ts
/** THÔNG TIN LỌC DOANH THU */
export default class ReportRevenueCriterial {
  fromDate: string = null;
  toDate: string = null;
  staffId: string = null;
  serviceId: string = null;
  roomId: string = null;
  medicalSupplyId: string = null;
  customerName: string = null;
  name: string = null;
  collectMoney: boolean = null;
  hasTreatment: boolean = null;
}